to int

+++ utils.py
@@ -72,7 +72,7 @@ def normalize_mask(mask, label_dict, resize=lambda img: img):
     mask = label_indices.reshape(tuple(mask.shape[:-1]) + (1,))
     mask = mask.transpose((2, 0, 1))
     mask = resize(mask)
-    mask = torch.from_numpy(mask)
+    mask = torch.from_numpy(mask).to(torch.int64)
     mask = torch.nn.functional.one_hot(mask, num_classes=Parameters.n_classes)
     mask = mask.permute(0, 3, 1, 2).squeeze(0).float()
     mask = torch.Tensor.numpy(mask)
